# specs/001-prompt-detect-untagged/contracts/api.yaml
openapi: 3.0.0
info:
  title: SBIR Transition Detection API
  description: API for detecting and retrieving evidence of untagged SBIR Phase III transitions.
  version: 1.0.0

paths:
  /detect:
    post:
      summary: Initiate a detection analysis
      description: >-
        Triggers an asynchronous analysis process for a given vendor or SBIR award
        to find potential transitions. Returns a task ID to poll for results.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vendor_identifier:
                  type: string
                  description: A unique vendor identifier (e.g., UEI or CAGE code).
                  example: "ABCDE12345"
                sbir_award_piid:
                  type: string
                  description: The PIID of a specific SBIR award to analyze.
                  example: "W911NF-18-C-0033"
              anyOf:
                - required: [vendor_identifier]
                - required: [sbir_award_piid]
      responses:
        '202':
          description: Analysis has been accepted and is in progress.
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    format: uuid
                    description: A unique ID for the analysis task.
                    example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        '400':
          description: Bad request, missing required identifier.

  /evidence/{id}:
    get:
      summary: Retrieve evidence for a detection
      description: Fetches the detailed evidence bundle for a specific detection ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the detection.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the evidence bundle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceBundle'
        '404':
          description: Detection not found.

components:
  schemas:
    EvidenceBundle:
      type: object
      properties:
        detection_id:
          type: string
          format: uuid
        likelihood_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        confidence:
          type: string
          enum: [High Confidence, Likely Transition]
        reason_string:
          type: string
          description: A human-readable explanation of the detection logic.
          example: "Sole-source IDV awarded 8 months after Phase II completion by the same agency."
        source_sbir_award:
          type: object
          properties:
            piid:
              type: string
            agency:
              type: string
            completion_date:
              type: string
              format: date
        source_contract:
          type: object
          properties:
            piid:
              type: string
            agency:
              type: string
            start_date:
              type: string
              format: date
            competition:
              type: object